% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_any_file}
\alias{download_any_file}
\title{Download any file related to the class}
\usage{
download_any_file(file_path_and_name, force_download = FALSE, mode = "wb")
}
\arguments{
\item{file_path_and_name}{The name of the path and file to download.}

\item{force_download}{Will download and overwrite an existing file if the
downloaded file has the same name.}

\item{mode}{A character indicating the mode with which to write the file.
Useful values are "w", "wb" (binary), "a" (append) and "ab". Not used for
methods "wget" and "curl". See also ‘Details’, notably about using "wb" for
Windows. See download.file() mode argument.}
}
\description{
This function downloads any file related to the class from the class GitHub
repository.
}
\examples{
 # Download an image from the class GitHub repository
 \dontrun{download_any_file("images/valentin.png")}

}
