% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_data}
\alias{download_data}
\title{Download data related to the class}
\usage{
download_data(file_name)
}
\arguments{
\item{file_name}{The name the file that has the data.}
}
\description{
This function downloads data related to the class from the class GitHub
repository.
}
\examples{
 # Download code from the first class
 \dontrun{download_data("test_data.rda")}

}
