% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_homework}
\alias{download_homework}
\title{Download a homework R Markdown file}
\usage{
download_homework(homework_number)
}
\arguments{
\item{homework_number}{A number indicating which homework number file to get.}
}
\description{
This function downloads homework files from the class GitHub repository.
}
\examples{
 # Download the first homework file
 \dontrun{download_homework(1)}

}
