% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_image}
\alias{download_image}
\title{Download an image related to the class}
\usage{
download_image(file_name, force_download = FALSE)
}
\arguments{
\item{file_name}{The name the image to download.}
}
\description{
This function downloads and image related to the class from the class GitHub
repository.
}
\examples{
 # Download an image from the class GitHub repository
 \dontrun{download_image("valentin.png")}

}
