% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_and_url_names.R
\name{get_version}
\alias{get_version}
\title{Get the package name and version number}
\usage{
get_version()
}
\description{
Returns the name of the package and the version number. This is useful to
check that one is using the most recent version of the package in case the
package is updated in the middle of the semester.
}
\examples{
 # Download the first homework file
 get_version()

}
