% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permCor_function.R
\name{permCor}
\alias{permCor}
\title{A permutation test and plotting function}
\usage{
permCor(x, y, n_samp = 10000, plot = T)
}
\arguments{
\item{x}{One continuous variable}

\item{y}{Another continuous variable}

\item{n_samp}{The number of permutation iterations. Defaults to 10000}

\item{plot}{Return a histogram of the permutation distribution with your sample correlation? Defaults to TRUE}
}
\description{
This function performs a permutation test for the correlation of two samples, returning the p-value associated with your original Pearson's correlation, relative to null permutation distribution.
}
\examples{
x <- 1:100
y <- jitter(x, factor=100)
permCor <- function(x, y, n_samp = 1000, plot = T)
}
\keyword{permutation}
