% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_package.R
\name{reinstall_package}
\alias{reinstall_package}
\title{Reinstalls the class package}
\usage{
reinstall_package()
}
\description{
This function reinstalls the class package. It is useful in case the package
has been updated in the middle of the semester.
}
\examples{
 # Reinstall the package
 \dontrun{reinstall_package()}

}
