% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_installed_packages.R
\name{update_installed_packages}
\alias{update_installed_packages}
\title{Installs packages that are used in the class}
\usage{
update_installed_packages(force_reinstall = FALSE)
}
\arguments{
\item{force_reinstall}{If this is set to TRUE, then any class packages that
have already been installed will be reinstalled. If this is FALSE, then only
packages that have not been installed yet will be installed.}
}
\description{
This function installs packages that are used in the class based on a list
of packages described in a file on the class GitHub repository.
}
\examples{
 # Install all the needed class packages
 \dontrun{update_installed_packages()}

}
