% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_class_code}
\alias{download_class_code}
\title{Download code related to the class}
\usage{
download_class_code(class_number, with_answers = FALSE)
}
\arguments{
\item{class_number}{A number indicating which class we should get the code
for.}

\item{with_answers}{A boolean indicating whether you want the class Markdown
document that contains all the code filled out, or you want the version
that will be used in class where we fill out the solutions as you go.}
}
\description{
This function downloads code that will be used in class from the class GitHub
repository.
}
\examples{
 # Download code from the first class
 \dontrun{download_class_code(1)}

}
